simulations/pyproject.toml
[project]
name = "neuromodulation-simulations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "k-wave-python>=0.4.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "torch>=2.10.0",
]
